+++ lib/reporting/ReportGenerator.js
@@ -3,7 +3,7 @@
  *
  * Created by Merlin Weemaes on 2/26/15.
  */
-var HtmlReporter = require('./HtmlReporter');
+var HtmlReporter = require('./html/HtmlReporter');
 var fs = require('fs');
 var reporterMapping = {'html': HtmlReporter};
 
@@ -16,5 +16,5 @@ exports.generate = function (opts, results) {
         console.log('report generator done');
     });
     report.writeReport(collector);
-}
+};
