+++ lib/reporting/html/templates/file.html
@@ -5,6 +5,6 @@
     <title></title>
 </head>
 <body>
-{{fileName}}
+{{source}}
 </body>
 </html>
