+++ lib/reporting/html/templates/index.html
